+++ src/main/java/org/threadly/concurrent/statistics/package-info.java
@@ -1,5 +1,7 @@
 
 /**
- * Package of tools and scheduler wrappers for getting statistics from concurrent systems.
+ * <p>Package of tools and scheduler wrappers for getting statistics from concurrent systems.</p>
+ * 
+ * <p>This package has been moved to {@link org.threadly.concurrent.wrapper.statistic}.</p>
  */
 package org.threadly.concurrent.statistics;
