+++ src/main/java/org/threadly/concurrent/wrapper/statistic/PrioritySchedulerStatisticTracker.java
@@ -1,26 +1,35 @@
-package org.threadly.concurrent.statistics;
+package org.threadly.concurrent.wrapper.statistic;
 
+import java.util.ArrayList;
+import java.util.Iterator;
+import java.util.List;
+import java.util.Map;
+import java.util.concurrent.Future;
 import java.util.concurrent.ThreadFactory;
 
+import org.threadly.concurrent.ConfigurableThreadFactory;
+import org.threadly.concurrent.PriorityScheduler;
 import org.threadly.concurrent.TaskPriority;
+import org.threadly.concurrent.collections.ConcurrentArrayList;
+import org.threadly.concurrent.wrapper.statistic.PriorityStatisticManager.TaskStatWrapper;
+import org.threadly.util.Clock;
+import org.threadly.util.Pair;
 
 /**
- * <p>An implementation of {@link org.threadly.concurrent.PriorityScheduler} which tracks run and 
- * usage statistics.  This is designed for testing and troubleshooting.  It has a little more 
- * overhead from the normal {@link org.threadly.concurrent.PriorityScheduler}.</p>
+ * <p>An implementation of {@link PriorityScheduler} which tracks run and usage statistics.  This 
+ * is designed for testing and troubleshooting.  It has a little more overhead from the normal 
+ * {@link PriorityScheduler}.</p>
  * 
  * <p>It helps give insight in how long tasks are running, how well the thread pool is being 
  * utilized, as well as execution frequency.</p>
  * 
- * @deprecated Moved to {@link org.threadly.concurrent.wrapper.statistic.PrioritySchedulerStatisticTracker}
- * 
  * @author jent - Mike Jensen
  * @since 4.5.0 (earlier forms existed since 1.0.0)
  */
-@Deprecated
-public class PrioritySchedulerStatisticTracker 
-                 extends org.threadly.concurrent.wrapper.statistic.PrioritySchedulerStatisticTracker 
-                 implements StatisticPriorityScheduler {
+public class PrioritySchedulerStatisticTracker extends PriorityScheduler 
+                                               implements StatisticPriorityScheduler {
+  protected final PriorityStatisticManager statsManager;
+  
   /**
    * Constructs a new thread pool, though no threads will be started till it accepts it's first 
    * request.  This constructs a default priority of high (which makes sense for most use cases).  
@@ -34,7 +43,8 @@ public class PrioritySchedulerStatisticTracker
    * @param poolSize Thread pool size that should be maintained
    */
   public PrioritySchedulerStatisticTracker(int poolSize) {
-    super(poolSize);
+    this(poolSize, DEFAULT_PRIORITY, 
+         DEFAULT_LOW_PRIORITY_MAX_WAIT_IN_MS, DEFAULT_NEW_THREADS_DAEMON);
   }
   
   /**
@@ -50,7 +60,7 @@ public class PrioritySchedulerStatisticTracker
    * @param useDaemonThreads {@code true} if newly created threads should be daemon
    */
   public PrioritySchedulerStatisticTracker(int poolSize, boolean useDaemonThreads) {
-    super(poolSize, useDaemonThreads);
+    this(poolSize, DEFAULT_PRIORITY, DEFAULT_LOW_PRIORITY_MAX_WAIT_IN_MS, useDaemonThreads);
   }
   
   /**
@@ -70,7 +80,7 @@ public class PrioritySchedulerStatisticTracker
    */
   public PrioritySchedulerStatisticTracker(int poolSize, TaskPriority defaultPriority, 
                                            long maxWaitForLowPriorityInMs) {
-    super(poolSize, defaultPriority, maxWaitForLowPriorityInMs);
+    this(poolSize, defaultPriority, maxWaitForLowPriorityInMs, DEFAULT_NEW_THREADS_DAEMON);
   }
   
   /**
@@ -92,7 +102,7 @@ public class PrioritySchedulerStatisticTracker
   public PrioritySchedulerStatisticTracker(int poolSize, TaskPriority defaultPriority, 
                                            long maxWaitForLowPriorityInMs, 
                                            boolean useDaemonThreads) {
-    super(poolSize, defaultPriority, maxWaitForLowPriorityInMs, useDaemonThreads);
+    this(poolSize, defaultPriority, maxWaitForLowPriorityInMs, useDaemonThreads, 1000);
   }
   
   /**
@@ -114,7 +124,7 @@ public class PrioritySchedulerStatisticTracker
   public PrioritySchedulerStatisticTracker(int poolSize, TaskPriority defaultPriority, 
                                            long maxWaitForLowPriorityInMs, 
                                            ThreadFactory threadFactory) {
-    super(poolSize, defaultPriority, maxWaitForLowPriorityInMs, threadFactory);
+    this(poolSize, defaultPriority, maxWaitForLowPriorityInMs, threadFactory, 1000);
   }
   
   /**
@@ -131,7 +141,8 @@ public class PrioritySchedulerStatisticTracker
    * @param maxStatisticWindowSize maximum number of samples to keep internally
    */
   public PrioritySchedulerStatisticTracker(int poolSize, int maxStatisticWindowSize) {
-    super(poolSize, maxStatisticWindowSize);
+    this(poolSize, DEFAULT_PRIORITY, 
+         DEFAULT_LOW_PRIORITY_MAX_WAIT_IN_MS, DEFAULT_NEW_THREADS_DAEMON, maxStatisticWindowSize);
   }
   
   /**
@@ -149,7 +160,8 @@ public class PrioritySchedulerStatisticTracker
    */
   public PrioritySchedulerStatisticTracker(int poolSize, boolean useDaemonThreads, 
                                            int maxStatisticWindowSize) {
-    super(poolSize, useDaemonThreads, maxStatisticWindowSize);
+    this(poolSize, DEFAULT_PRIORITY, DEFAULT_LOW_PRIORITY_MAX_WAIT_IN_MS, 
+         useDaemonThreads, maxStatisticWindowSize);
   }
   
   /**
@@ -171,7 +183,8 @@ public class PrioritySchedulerStatisticTracker
   public PrioritySchedulerStatisticTracker(int poolSize, TaskPriority defaultPriority, 
                                            long maxWaitForLowPriorityInMs, 
                                            int maxStatisticWindowSize) {
-    super(poolSize, defaultPriority, maxWaitForLowPriorityInMs, maxStatisticWindowSize);
+    this(poolSize, defaultPriority, maxWaitForLowPriorityInMs, 
+         DEFAULT_NEW_THREADS_DAEMON, maxStatisticWindowSize);
   }
   
   /**
@@ -194,8 +207,8 @@ public class PrioritySchedulerStatisticTracker
   public PrioritySchedulerStatisticTracker(int poolSize, TaskPriority defaultPriority, 
                                            long maxWaitForLowPriorityInMs, 
                                            boolean useDaemonThreads, int maxStatisticWindowSize) {
-    super(poolSize, defaultPriority, maxWaitForLowPriorityInMs, useDaemonThreads, 
-          maxStatisticWindowSize);
+    this(poolSize, defaultPriority, maxWaitForLowPriorityInMs, useDaemonThreads, 
+         maxStatisticWindowSize, false);
   }
   
   /**
@@ -219,8 +232,8 @@ public class PrioritySchedulerStatisticTracker
                                            long maxWaitForLowPriorityInMs, 
                                            ThreadFactory threadFactory, 
                                            int maxStatisticWindowSize) {
-    super(poolSize, defaultPriority, maxWaitForLowPriorityInMs, 
-          threadFactory, maxStatisticWindowSize);
+    this(poolSize, defaultPriority, maxWaitForLowPriorityInMs, 
+         threadFactory, maxStatisticWindowSize, false);
   }
 
   /**
@@ -235,7 +248,8 @@ public class PrioritySchedulerStatisticTracker
    */
   public PrioritySchedulerStatisticTracker(int poolSize, 
                                            int maxStatisticWindowSize, boolean accurateTime) {
-    super(poolSize, maxStatisticWindowSize, accurateTime);
+    this(poolSize, DEFAULT_PRIORITY, DEFAULT_LOW_PRIORITY_MAX_WAIT_IN_MS, 
+         DEFAULT_NEW_THREADS_DAEMON, maxStatisticWindowSize, accurateTime);
   }
   
   /**
@@ -250,7 +264,8 @@ public class PrioritySchedulerStatisticTracker
    */
   public PrioritySchedulerStatisticTracker(int poolSize, boolean useDaemonThreads, 
                                            int maxStatisticWindowSize, boolean accurateTime) {
-    super(poolSize, useDaemonThreads, maxStatisticWindowSize, accurateTime);
+    this(poolSize, DEFAULT_PRIORITY, DEFAULT_LOW_PRIORITY_MAX_WAIT_IN_MS, 
+         useDaemonThreads, maxStatisticWindowSize, accurateTime);
   }
   
   /**
@@ -269,8 +284,8 @@ public class PrioritySchedulerStatisticTracker
   public PrioritySchedulerStatisticTracker(int poolSize, TaskPriority defaultPriority, 
                                            long maxWaitForLowPriorityInMs, 
                                            int maxStatisticWindowSize, boolean accurateTime) {
-    super(poolSize, defaultPriority, maxWaitForLowPriorityInMs, 
-         maxStatisticWindowSize, accurateTime);
+    this(poolSize, defaultPriority, maxWaitForLowPriorityInMs, 
+         DEFAULT_NEW_THREADS_DAEMON, maxStatisticWindowSize, accurateTime);
   }
   
   /**
@@ -290,8 +305,10 @@ public class PrioritySchedulerStatisticTracker
   public PrioritySchedulerStatisticTracker(int poolSize, TaskPriority defaultPriority, 
                                            long maxWaitForLowPriorityInMs, boolean useDaemonThreads, 
                                            int maxStatisticWindowSize, boolean accurateTime) {
-    super(poolSize, defaultPriority, maxWaitForLowPriorityInMs, useDaemonThreads, 
-          maxStatisticWindowSize, accurateTime);
+    this(poolSize, defaultPriority, maxWaitForLowPriorityInMs, 
+         new ConfigurableThreadFactory(PrioritySchedulerStatisticTracker.class.getSimpleName() + "-", 
+                                       true, useDaemonThreads, Thread.NORM_PRIORITY, null, null), 
+         maxStatisticWindowSize, accurateTime);
   }
   
   /**
@@ -312,7 +329,197 @@ public class PrioritySchedulerStatisticTracker
                                            long maxWaitForLowPriorityInMs, 
                                            ThreadFactory threadFactory, 
                                            int maxStatisticWindowSize, boolean accurateTime) {
-    super(poolSize, defaultPriority, maxWaitForLowPriorityInMs, threadFactory, 
-          maxStatisticWindowSize, accurateTime);
+    super(new StatisticWorkerPool(threadFactory, poolSize, 
+                                  new PriorityStatisticManager(maxStatisticWindowSize, accurateTime)), 
+          maxWaitForLowPriorityInMs, defaultPriority);
+    
+    this.statsManager = ((StatisticWorkerPool)workerPool).statsManager;
+  }
+  
+  @Override
+  public List<Runnable> shutdownNow() {
+    // we must unwrap our statistic tracker runnables
+    List<Runnable> wrappedRunnables = super.shutdownNow();
+    List<Runnable> result = new ArrayList<Runnable>(wrappedRunnables.size());
+    
+    Iterator<Runnable> it = wrappedRunnables.iterator();
+    while (it.hasNext()) {
+      Runnable r = it.next();
+      if (r instanceof TaskStatWrapper) {
+        TaskStatWrapper tw = (TaskStatWrapper)r;
+        if (! (tw.task instanceof Future) || ! ((Future<?>)tw.task).isCancelled()) {
+          result.add(tw.task);
+        }
+      } else {
+        // this typically happens in unit tests, but could happen by an extending class
+        result.add(r);
+      }
+    }
+    
+    return result;
+  }
+  
+  /**
+   * Wraps the provided task in our statistic wrapper.  If the task is {@code null}, this will 
+   * return {@code null} so that the parent class can do error checking.
+   * 
+   * @param task Runnable to wrap
+   * @param priority Priority for runnable to execute
+   * @return Runnable which is our wrapped implementation
+   */
+  private Runnable wrap(Runnable task, TaskPriority priority) {
+    if (priority == null) {
+      priority = getDefaultPriority();
+    }
+    if (task == null) {
+      return null;
+    } else {
+      return new TaskStatWrapper(statsManager, priority, task);
+    }
+  }
+
+  @Override
+  protected OneTimeTaskWrapper doSchedule(Runnable task, long delayInMillis, TaskPriority priority) {
+    return super.doSchedule(new TaskStatWrapper(statsManager, priority, task), 
+                            delayInMillis, priority);
+  }
+
+  @Override
+  public void scheduleWithFixedDelay(Runnable task, long initialDelay,
+                                     long recurringDelay, TaskPriority priority) {
+    super.scheduleWithFixedDelay(wrap(task, priority), initialDelay, recurringDelay, priority);
+  }
+
+  @Override
+  public void scheduleAtFixedRate(Runnable task, long initialDelay,
+                                  long period, TaskPriority priority) {
+    super.scheduleAtFixedRate(wrap(task, priority), initialDelay, period, priority);
+  }
+
+  @Override
+  public List<Long> getExecutionDelaySamples() {
+    return statsManager.getExecutionDelaySamples();
+  }
+  
+  @Override
+  public List<Long> getExecutionDelaySamples(TaskPriority priority) {
+    return statsManager.getExecutionDelaySamples(priority);
+  }
+
+  @Override
+  public double getAverageExecutionDelay() {
+    return statsManager.getAverageExecutionDelay();
+  }
+
+  @Override
+  public double getAverageExecutionDelay(TaskPriority priority) {
+    return statsManager.getAverageExecutionDelay(priority);
+  }
+
+  @Override
+  public Map<Double, Long> getExecutionDelayPercentiles(double... percentiles) {
+    return statsManager.getExecutionDelayPercentiles(percentiles);
+  }
+
+  @Override
+  public Map<Double, Long> getExecutionDelayPercentiles(TaskPriority priority, 
+                                                        double... percentiles) {
+    return statsManager.getExecutionDelayPercentiles(priority, percentiles);
+  }
+
+  @Override
+  public List<Long> getExecutionDurationSamples() {
+    return statsManager.getExecutionDurationSamples();
+  }
+
+  @Override
+  public List<Long> getExecutionDurationSamples(TaskPriority priority) {
+    return statsManager.getExecutionDurationSamples(priority);
+  }
+
+  @Override
+  public double getAverageExecutionDuration() {
+    return statsManager.getAverageExecutionDuration();
+  }
+
+  @Override
+  public double getAverageExecutionDuration(TaskPriority priority) {
+    return statsManager.getAverageExecutionDuration(priority);
+  }
+
+  @Override
+  public Map<Double, Long> getExecutionDurationPercentiles(double... percentiles) {
+    return statsManager.getExecutionDurationPercentiles(percentiles);
+  }
+
+  @Override
+  public Map<Double, Long> getExecutionDurationPercentiles(TaskPriority priority, 
+                                                           double... percentiles) {
+    return statsManager.getExecutionDurationPercentiles(priority, percentiles);
+  }
+
+  @Override
+  public List<Pair<Runnable, StackTraceElement[]>> getLongRunningTasks(long durationLimitMillis) {
+    return statsManager.getLongRunningTasks(durationLimitMillis);
+  }
+
+  @Override
+  public int getLongRunningTasksQty(long durationLimitMillis) {
+    return statsManager.getLongRunningTasksQty(durationLimitMillis);
+  }
+  
+  @Override
+  public void resetCollectedStats() {
+    statsManager.resetCollectedStats();
+  }
+  
+  @Override
+  public long getTotalExecutionCount() {
+    return statsManager.getTotalExecutionCount();
+  }
+
+  @Override
+  public long getTotalExecutionCount(TaskPriority priority) {
+    return statsManager.getTotalExecutionCount(priority);
+  }
+  
+  /**
+   * <p>An extending class of {@link WorkerPool}, allowing us to gather statistics about how 
+   * workers are used in the executor.  An example of such statistics are how long tasks are 
+   * delayed from their desired execution.  Another example is how often a worker can be reused vs 
+   * how often they have to be created.</p>
+   * 
+   * @author jent - Mike Jensen
+   * @since 4.5.0
+   */
+  protected static class StatisticWorkerPool extends WorkerPool {
+    protected final PriorityStatisticManager statsManager;
+  
+    protected StatisticWorkerPool(ThreadFactory threadFactory, int poolSize, 
+                                  PriorityStatisticManager statsManager) {
+      super(threadFactory, poolSize);
+      
+      this.statsManager = statsManager;
+    }
+    
+    @Override
+    public TaskWrapper workerIdle(Worker worker) {
+      TaskWrapper result = super.workerIdle(worker);
+
+      // may not be a wrapper for internal tasks like shutdown
+      if (result != null && result.getContainedRunnable() instanceof TaskStatWrapper) {
+        long taskDelay = Clock.lastKnownForwardProgressingMillis() - result.getPureRunTime();
+        TaskStatWrapper statWrapper = (TaskStatWrapper)result.getContainedRunnable();
+        ConcurrentArrayList<Long> priorityStats = 
+            statsManager.getExecutionDelaySamplesInternal(statWrapper.priority);
+  
+        synchronized (priorityStats.getModificationLock()) {
+          priorityStats.add(taskDelay);
+          statsManager.trimWindow(priorityStats);
+        }
+      }
+      
+      return result;
+    }
   }
 }
