+++ src/main/java/org/threadly/concurrent/wrapper/statistic/PriorityStatisticManager.java
@@ -1,4 +1,4 @@
-package org.threadly.concurrent.statistics;
+package org.threadly.concurrent.wrapper.statistic;
 
 import java.util.ArrayList;
 import java.util.Deque;
