+++ src/test/java/org/threadly/concurrent/wrapper/statistic/NoThreadSchedulerStatisticTrackerTest.java
@@ -1,4 +1,4 @@
-package org.threadly.concurrent.statistics;
+package org.threadly.concurrent.wrapper.statistic;
 
 import static org.junit.Assert.*;
 import static org.threadly.TestConstants.*;
@@ -13,7 +13,7 @@ import org.threadly.concurrent.TaskPriority;
 import org.threadly.test.concurrent.TestRunnable;
 import org.threadly.test.concurrent.TestUtils;
 
-@SuppressWarnings({"javadoc", "deprecation"})
+@SuppressWarnings("javadoc")
 public class NoThreadSchedulerStatisticTrackerTest extends NoThreadSchedulerTest {
   @Before
   @Override
