+++ src/test/java/org/threadly/util/debug/ProfilerTest.java
@@ -14,9 +14,9 @@ import java.util.concurrent.TimeoutException;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
-import org.threadly.concurrent.statistics.PrioritySchedulerStatisticTracker;
 import org.threadly.concurrent.StrictPriorityScheduler;
 import org.threadly.concurrent.future.ListenableFuture;
+import org.threadly.concurrent.wrapper.statistic.PrioritySchedulerStatisticTracker;
 import org.threadly.test.concurrent.TestCondition;
 import org.threadly.util.Clock;
